fix: first step to fully implement variable byte integers, fixed header remaining length increased from u16 to u32

+++ mqtt/src/control_packets/mqtt_auth/auth.rs
@@ -12,7 +12,7 @@ pub struct Auth {
 }
 
 impl Serialization for Auth {
-    fn read_from(stream: &mut dyn Read, remaining_length: u16) -> Result<Auth, Error> {
+    fn read_from(stream: &mut dyn Read, remaining_length: u32) -> Result<Auth, Error> {
         let mut aux_buffer = vec![0; remaining_length as usize];
         stream.read_exact(&mut aux_buffer)?;
         let mut buffer = aux_buffer.as_slice();
