+++ mqtt/src/control_packets/mqtt_connect/connect.rs
@@ -100,7 +100,7 @@ pub struct Connect {
 }
 
 impl Serialization for Connect {
-    fn read_from(stream: &mut dyn Read, remaining_length: u16) -> Result<Connect, Error> {
+    fn read_from(stream: &mut dyn Read, remaining_length: u32) -> Result<Connect, Error> {
         let mut aux_buffer = vec![0; remaining_length as usize];
         stream.read_exact(&mut aux_buffer)?;
         let mut buffer = aux_buffer.as_slice();
