+++ mqtt/src/control_packets/mqtt_puback/puback.rs
@@ -71,7 +71,7 @@ pub struct Puback {
 }
 
 impl Serialization for Puback {
-    fn read_from(stream: &mut dyn Read, remaining_length: u16) -> Result<Self, std::io::Error> {
+    fn read_from(stream: &mut dyn Read, remaining_length: u32) -> Result<Self, std::io::Error> {
         let mut aux_buffer = vec![0; remaining_length as usize];
         stream.read_exact(&mut aux_buffer)?;
         let mut buffer = aux_buffer.as_slice();
