+++ mqtt/src/control_packets/mqtt_subscribe/subscribe.rs
@@ -58,7 +58,7 @@ pub struct Subscribe {
 }
 
 impl Serialization for Subscribe {
-    fn read_from(stream: &mut dyn Read, remaining_length: u16) -> Result<Self, Error> {
+    fn read_from(stream: &mut dyn Read, remaining_length: u32) -> Result<Self, Error> {
         let mut aux_buffer = vec![0; remaining_length as usize];
         stream.read_exact(&mut aux_buffer)?;
         let mut buffer = aux_buffer.as_slice();
